Verifier implementation using an external service (based on #122, supersedes #223)
This implementation calls out the external Microservice ipaas-verifier for doing the verification of a connector and has the following advantages over the locally process forking implementation of #223:

* Much faster because of an already 'prewarmed' running ipaas-verifier service
* Does not eat up memory and CPU resources by starting external Maven processes which are also quite network intensive since a lot of dependencies needs to be downloaded
* No local process management required like proper
* Simpler implementation
* Doesn't need to call an external CLI command
* Standard REST / Microservice technology
* Supports a the full set of data that a Camel ComponentVerifier can return, in addition to `code` and `description`
  - Parameters which are affected by an error
  - Attributes holding additional error information (i.e. whether a parameter is missing or an unexpected parameter has been send.
 * Uses always both scopes: First verify the PARAMETERS before checking the CONNECTIVITY. Only if the initial parameter validation succeeds, a the real connection is called. This result in much faster responses if the given parameter set is unsound.
 * Does not add a GAV to the connector, which is factual _wrong_. GAVs are connected to Actions, a connector might reference an Action as its 'default verification action', but must not duplicate the (possibly contradicting) GAV of an action. Currently this association happens on the ipaas-verifier side. This had to be changed anyway.
 * Supports twitter, salesforce and http instead of only twitter (however could be easily added).

 The disadvantage of this approach is, that it is currently much simpler and supports three connectors: twitter, salesforce and http which are pre-started in the service. There is no dynamic loading of additional connectors, since this is not in the scope of the current Sprint. The setup is however expandable to deal with dynamic connectors, too.

 ---

The `LocalProcessVerifier` code is still present as reference, however due to changes in the model does not fully work and needs to be adapted (e.g. not to use a connector GAV but an action GAV and adapt to the interface of the Verifier). However I doubt that we should add efforts into this for now. Actually I would recommend to remove the code as it has served its purpose of a POC well but might not be needed anymore right now.

The templates has been adapted, the image builds has been setup and the Twitter validation has been tested end-to-end (starting from ipaas-rest, ipass-ui integration is still pending). For testing the Salesforce verification I would need proper credentials for it.