#27 rest connector for Trades API
This creates a connector for a single operation on the Trades API REST
resource (day-trade-api[1]), this is done by delegating to Camel REST
component that in turn uses Camel Undertow component to perform the
actual interaction. Camel Jackson component is used to unmarshal the
response received from the service.

This has support for only one REST operation (getTrade[2]), but can be
used as a template to create connectors for other operations.

To initiate the exchange, the incoming message needs to have `orderId`
header set to the ID of the order that needs to be fetched.

The Java class `org.example.ipaas.Trade` has been vendored within the
connector as `com.redhat.ipaas.connector.daytrade.Trade`.

[1] https://github.com/kcbabo/ipaas-demo/tree/master/day-trade-api
[2] https://github.com/kcbabo/ipaas-demo/blob/master/day-trade-api/src/main/java/org/example/ipaas/OrderDesk.java#L33
[3] https://github.com/kcbabo/ipaas-demo/blob/master/day-trade-api/src/main/java/org/example/ipaas/Trade.java