Fixed problem with multiple checks for the same integration
This is uses a state map which is checked when processing a 'change-event' so
that no second job is triggered if there is already one.

Also added some logging and clarified the various steps happening (these could be even fine granular integration step states)

However there is still the problem with expiring tokens. Because of the
reordering, a GitHub API will fail first, whose error message is a bit obfuscated.

Would be awesome if there would be a check for verifying whether a token is till valid.