Introduced "steps" within an state handler.
Steps can be used to mark certain milestones until reaching a target state.
Each step which was done sets a mark (label) in the integration so that
it doesn't need to be redone when the state handler is retried.

This is based on #349 but uses a list of labels for the performed steps instead of a single numeric index. This is the advantage that the order can be easily changed withoud having to upgrade existing integration and it is easier to see by looking at the domain objects what kind of actions has been already successully be performed. The current labels are "github-setup" and "openshift-setup".