fix: Can't Verify Saleforce Connection (#12)
Removes the `operationName` added by `SalesforceVerifier` to the
parameters map. The `operationName` is not required and I think it
should not be added.

Also changes the connector name from `salesforce-upsert-contact` to just
`salesforce`.

Also creates Camel components without starting them as it is not needed
for verifying the component that the component is started.

Assumes:
 - component does not need to be started for verification to succeed
 - can use _mother_ component (i.e. salesforce) instead of connector (
   i.e _salesforce-upsert-contact_)

Does not:
 - make the same change for other verifiers (HTTP, Twitter)