fix(credentials): Finishing OAuth credential fl...
...ow on a wrong URL (#431)

When the UI specifies the return URL in credential acquisition it should
be taken as an absolute path not relative path to the API base URI.

This changes to compute the absolute URI at the point the UI submits the
relative URI to return to on successful OAuth acquisition.

I've also taken the opportunity to declare Mockito as _test_ dependency.