Use field level help icons instead of placeholders to display help text
Pattern here http://www.patternfly.org/pattern-library/forms-and-controls/field-level-help/

This is just to check how viable is a quick fix with ng2-dynamic-forms. Since we -probably- need tooltips we need to use ngx-bootstrap Tooltip feature,
which can not be injected as innerHTML being it a Directive.
Next try is to write a custom form control and address this properly (https://blog.thoughtram.io/angular/2016/07/27/custom-form-controls-in-angular-2.html).