feat(rest): Parameter validation (#440)
This configures RESTEasy to validate parameters in accordance to JSR 303
(Bean Validation). It also adds JSR 303 annotations to interfaces in
`io.syndesis.rest.v1.operations` package so that the standard CRUD
method parameters and to other *Handler implementations so that they are
all validated.

A new ExceptionMapper was added to provide JSON output in case of
constraint validation.

With the specific way the exceptions are handled in RESTEasy
(handler for [Reader|Writter]Exception is not looked up the inheritance
chain) two ExceptionMappers to handle those are added.

And with those changes in the exception handling it made sense to create
a base class for ExceptionMappers.

The only way to for the validations to work with Immutables is to use
the standard Java Beans method signatures (get/set) as there is
currently no way to specify an annotation on the field with
Immutables[1].

There is still more work to be done on validations, for instance deeper
syntax validations might be beneficial (e.g. disallow certain characters
in IDs to prevent injection).

fixes #440

[1] https://github.com/immutables/immutables/issues/475