feat(development): A faster dev turnaround (#445)
This adds two Maven profiles `flash` that disables all but necessary
Maven plugins by excluding them from the build, and `deploy` that uses
fabric8-maven-plugin to deploy to OpenShift (tested with minishift).

Both are activated by the presence of `deploy` property, so the easiest
way to run them both is:

    $ ./mvnw -Ddeploy

Which should be the same as:

    $ ./mvnw -Pflash,deploy

On my machine this takes ~21 sec.

fixes #445