fix(credentials): Determinisim in CredentialPro...
...vider registration (#447)

The way we register CredentialProviders is not deterministic and can
fail to register all CredentialProviders configured in @Configuration
classes that are incorporated in the Spring Application Context after
CredentialsConfiguration @Configuration class is done configuring and
registering them.

This reverses the role from "I'll find you" to "I'll register with you"
for CredentialProvider @Configurations. It also resurects the Configurer
classes so that CredentialProviders don't end up being singletons in the
Spring Application Context. They should not be accessible by other means
other than through CredentialProviderLocator and it alone should handle
their lifecycle.

fixes #447