feat(client side state): Implementation of clie...
...nt side state

OAuth1 and OAuth2 implementations used different state storage, this
changes that to use the same storage, which has been implemented on the
client side as a cookie.

For this the flow state had to be split into two variants one for OAuth1
and one for OAuth2, and state access had to be revised thought the
implementation.

This implements persisting state on the client side via secure cookies
as defined in RFC6896[1].

This is basis for storing OAuth flow state in #457.

The configuration needed for this to function can be specified via
environment variables:

    CLIENT_STATE_AUTHENTICATION_ALGORITHM _(e.g. HmacSHA1)_
    CLIENT_STATE_AUTHENTICATION_KEY _(e.g. random 20 bytes)_
    CLIENT_STATE_ENCRYPTION_ALGORITHM _(e.g. AES/CBC/PKCS5Padding)_
    CLIENT_STATE_ENCRYPTION_KEY _(e.g. random 20 bytes)_
    CLIENT_STATE_TID _(e.g. 1)_

If configuration is not defined random values are generated, wich will
issue a warning as random values in pod instances will not failover or
work after restart as the keys would differ.

Also, separated OAuth1 and OAuth2 logic, it is now pushed in the
`CredentialProvider`s so no more iffy logic in `Credentials`, yay!

[1] https://tools.ietf.org/html/rfc6896

Fixes #457