fix(project): Dependency issues
This reinstates the dependency plugin checks. Also of note fixes
incorrect version of jackson-annotations (resolved 2.8.0, rest of the
jackson dependencies are at 2.8.8).

A number of jersey dependencies were excluded as they are not needed.

`syndesis-maven-plugin` module was placed before the `runtime` module
in order to be built in proper order.