chore(configuration): Proxy API and Keycloak, t...
...urn on TLS

This configures the `yarn start` (in turn `ng serve`) to proxy the API
(syndesis-rest and datamaper) via `/api` endpoint and Keycloak via
`/auth` endpoint. This enables the UI and the backend to reside on the
same host domain which in turn allows setting Cookies and simplifies
CORS.
It also enables TLS (`https`) with self-signed certificate. That enables
setting secure Cookies.
Since the request travels through two reverse proxies (the middleware
proxy setup by `ng serve` and HA Proxy configured by OpenShift router)
proxy adds an additional HTTP header to each request
(`X-Forwarded-Origin`) that the backend can use to determine the
original hostname, as the `Forwarded` or `X-Forwarded-For` headers will
be overwritten by HA proxy and that information would be lost.