refactor(rule-filter): Adapt data format of a basic filter step according to the UI needs.
It's an ugly change. sorry. Instead of the typed interface
configuredProperties is used which contains values as serialized JSON strings.

Its change that way to make it happen and serious efforts should be undertaken to come to a more typed interface.

For possible options see https://github.com/syndesisio/syndesis-rest/issues/451#issuecomment-321249672