fix(oauth): Fixes cookie path and JSON deserial...
...ization

Better definition of the client side state cookie with the Path and
and Secure parameters set -- but without HttpOnly flag so that the
JavaScript can set the cookie, the cookie should traverse back and forth
when set from UI and received on OAuth callback.

Tries its best to remove all cookies that are not needed:
 - cookies that are stale
 - cookies that are from another edition
 - cookies that are no longer needed

Adds Jackson module specific to
`org.springframework.social.oauth1.OAuthToken` deserialization.

Fixes #503