feat(credentials): Mark connections as derived
This adds a flag `derived` to Connections, derived Connections are those
that did not get their `configuredProperties` from user input but from
other means, such as OAuth dance performed by Credentials. Credentials
will also set this flag to `true` for Connections that pass successfully
through the OAuth acquisition flow and are saved at that point.

Once set the flag remains set until the UI chooses to reset it to
`false`, which it might if we ever decide to offer editing derived
Connections.

Fixes #513