feat(validation): New resource for validation
Adds a new connection sub-resource (`validation`) that can be used to
perform validation of the soon-to-be created connection resource.

Full endpoint of the resource is `/api/v1/connections/validation` and
for connection payloads sent via `POST` method it will respond with
status `204` and 0-length body if the connection passes validation; or
status `400` and body containing a set of objects containing `property`
(what property) and `message` (validation message) properties.

Fixes #475