fix(oauth): Adds more detailed status of OAuth ...
...callback in the fragment

This would help the UI know if OAuth authorization succeeded or not. The
fragment of the given `returnUrl` is now ignored and replaced with a URL
encoded JSON that resembles something like:

    #%7B%22connectorId%22:%22test-provider%22,%22message%22:%22Successfully%20authorized%20Syndesis's%20access%22,%22status%22:%22SUCCESS%22%7D

(for successful authorization). Which translates to the following JSON:

    {
      "connectorId: "test-provider", // can be "twitter" or "salesforce"
      "message": "Successfully authorized Syndesis's access",
      "status": "SUCCESS"
    }

Status can also be "FAILURE" indication, well, failure in authorization,
the message offers a bit more detail.

Note(1): connectorId JSON property can be `null`, for instance if the
callback was invoked without `cred-` cookies, in that case there is no
flow to look for connectorId.

Note(2): The redirect might lead to the root of the application (i.e to
`/`) for the same reason.

Fixes #514