fix(validation): Don't detect updated unique pr...
...operties as non unique

The unique validation would look for a presence of a value and solely
based on that would declare it unique or non-unique. The problem with
that is, of course, that updates no longer work as it would detect it's
own value in the database and declare it non-unique. So one could create
but not update without changing the name.
This changes the lookup to return all IDs that have that unique property
with the specified value, and then if the ID of the object is contained
in those will not declare that value as non-unique.

Fixes #552