chore(metadata): Adds support for dynamic actio...
...n properties

Adds /api/v1/connections/{connectionId}/actions/{actionId}/properties
endpoint that responds to GET request and in the query parameters
expects any action property - value pairs that have been selected by the
user and offers suggestions as to values of other action properties.

The implementation delegates to syndesis-verifier and the endpoint
created in syndesision/syndesis-verifier#18[1] that in turn delegates
to Camel Metadata extension.

[1] syndesisio/syndesis-verifier#18