fix(settings): Allows updating OAuth settings w...
...ith empty or null values

In the effort to remove OAuth configuration OAuth settings endpoint must
allow empty or null values. With trim-to-null empty values were
transformed to null and this did not allow the Connector's configured
properties to be updated. This builds on the PR#572 and allows setting
empty values for OAuth properties in the effort to remove those from
`configuredProperties`.

Fixes #536