chore(metadata): Fetch action properties and da...
..ta shapes

Adds a new endpoint to fetch both action properties and input/output
data shapes dynamically.

Changes the `MetadataAdapter` contact to return a `SyndesisMetadata`
that holds the dynamic action properties and input/output data shapes (
of certain type).

Refactored the `SalesforceMetadataAdapter` to the new contract and to
the changes made in Camel on how the Salesforce `MetadataExtension`
returns data for the global objects metadata call - it used to be custom
JSON payload, the newer edition returns JSON schema payload, akin to
what the describe object call returns. The adapter is compatible with
both payload types (JSON schema and custom JSON) to preserve
compatibility with Camel versions.

The existing endpoint is left to be backward compatible. Will be removed
with the corresponding change in the `syndesis-rest`.

Fixes #21