fix(metadata): Better `null` handling
Verifier has no knowledge of the properties required for a Syndesis
action. So `syndesis-rest` sends `null` values for properties action
requires but the user has not specified yet. This is not helpful for
Camel metadata extension, only as a hint to the `MetadataAdapter`s to
set those action properties if requested.

Also SalesforceMetadataAdapter should not fetch schema/data shape for
unspecified (`null`) Salesforce object names.