chore: Define new connectors
This adds Salesforce and SQL stored procedure connectors needed for TP1.
Removes Salesforce and Twitter actions that do not have an
implementation.
Adds data shape definition to Twitter mention connector.
It also removes the `defaultValue` for Salesforce connection's
`instanceId` property that was getting set on new connections and would
cause issues connection to salesforce instance (was set to
_xx.salesforce.com_).

Fixes #589