refactor(deployment): Changed from "Rolling" to "Recreate" strategy
The issue with a 'Rolling' update is, that for some time it needs two containers running in parallel.
However, for resource restricted setup likes on OpenShift Online this very likely causes to go over the limits (of 4 GB).

Therefor changed back to a recreate strategy to avoid this extra memory costs.