chore: Add Salesforce streaming connectors
This adds thee connectors:
 - salesforce-on-create
 - salesforce-on-update
 - salesforce-on-delete

These connectors are intended to be used to start the integration, i.e.
Camel `from` endpoints.

In addition to authentication properties they have one property open for
configuration `sObjectName` the name of the object to get events for.

Internally connectors generate `topicName` and `query` endpoint
properties and remove the `sObjectName` property to prevent class lookup
by the given object name that would lead to IllegalArgumentException.

Output of the connector is `AnyObject` DTO that contains a single `Id`
property with the identifier of the object the event was generated for.