chore: Sync connectors with syndesis-rest
This updates the `deployment.json` by removing the connectors that are
only for demo purposes (that is either do not work anymore or do not
have an implementation) and updating connector/action definitions to
match those in connectors.

Adds a test to make sure that the connectors defined in
`deployment.json` can be resolved and match in definition.