feat(setup): Add setup JAX-RS application for first time user setup
The setup application provides a single endpoint, at `/api/setup`. This is a one
time only handler: once it is updated, it is unusable. The endpoint accepts the
following methods:

`GET`: This checks if setup is required. It returns `204` (`No Content`) if the
setup handler is still enabled, and `410` (`Gone`) if it is already
configured and hence can't be called again.

`PUT`: This attempts to update the configuration as requested. This
handler will return `410` (`Gone`) if configuration is no longer
required. If the endpoint is still enabled, then the handler accepts a
JSON body with the following format:

```json
{
  "gitHubOAuthConfiguration": {
    "clientId": "yourclientid",
    "clientSecret": "topsecret"
  }
}
```

The format of this is extensible for any future first time user setup
steps.