fix: work with options specified as parameters (2)
The fix in 60c71820bd664c83c413fed1755a016268e16c26 would override
component properties which would result in multiple endpoints racing to
set a value in `options`.

This uses URI of the next processor (to-endpoint) and parses the
parameter from that. If the parameter is not defined it fallbacks to the
value in component configuration.