feat(runtime): Prefetch connector catalog
When creating project files, a `CamelCatalog` and `CamelConnectorCatalog`
is required to calculate the proper endpoint URL

Up to now, this is only done via Grape which does a live
introspection and download of connector jars.

A new plugin goal `syndesis:extract-connector-descriptor` has been
introduced to fetch this information offline.
By adding this to `META-INF/camel/camel-meta.json`, the runtime
can query this file and build up a CamelCatalog from this information only.
This avoids to reach out to the internet, which can have unexpected
side effects related to the performance.