fix: if the status of the integration revision ...
...does not change update the last revision

Currently we add a new integration revision for every status check,
about one every minute, this leads to a huge number of rows being added
to the database.

This change updates the last integration revision if the status of the
last revision does not change with the update. This way we have the
number of revisions tied to the number of state changes.

Fixes #638