fix(verifier): add hook for verifier extension ...
...lookup

Camel extensions are contained in a list of suppliers, connectors
components lookup VerifierExtension as a part of the initialization,
this can add a UNSUPPORTED extension (i.e. one that returns that
verification is not supported) for connectors that are derived from
components that do not provide a VerifierExtension.

In the case of SQL stored procedure connector we register extensions
through the connector itself.

This provides a constructor variant that accepts explicit extension
implementation class for looking up the verifier extension so we can ask
for specific implementation of the SQL stored procedure connector
extension and not get the UNSUPPORTED extension instead.