refactor(git): Remove Git
This is the start of removing calling out to GitHub for storing integrations.
Instead the generated runtime files are directly send as binary input
to a s2i source build. These files exist only temporarily on disc.

The first step was to remove git and switch to calling a binary build.
It is currently not yet end2end tested.

The next steps are also to get rid of all Keycloak references and adapt
the deployment accordingly.

Still need to understand how the OpenShift oauthclient can be used for
substituting authentication (can it ?)