refactor: Use streaming instead of temp files
The data generated is now written directly into the stream given to the OpenShift build.
The files are still generated completely in memory as the tar archiver
from commons-compress only works with files or byte arrays (but not streams ;-(

However, since the files are supposed to be smalle (syndesis.yml, datamapper.json, ...) this shoudl be no problem.