fix: Proper build and deployment handling
For now we are triggering a deployment explicitly, to circumvent
issue on our target platform which has problems with deployment triggers.

However I was not able to disable image triggers on the deployment, otherwise
imagestream could not be used. This cause a double deployment on
platforms with working triggers
This is not nice but should not harm to much.