fix: load data shapes when datamapping step is ...
...initialized

When the datamapper is initialized it needs to be provided with proper
data shapes. Datashapes are a) dynamic and b) sometimes not loaded.
The later (b) is when metadata is fetched for multi-step action on the
next-to-last step (as metadata suggestions are not needed for the last
step), or when configuring a single step action (so metadata suggestion
is fetched without specifying the configured action properties).

This adds a call to fetch metadata and set the input/output data shapes
for the start and end steps, which is limiting and does not handle
errors.

Fetching of the metadata could also be performed in parallel as
data shapes of actions should be independent of each other.

Error handling should be performed.

Fetching metadata should be done for other steps (not just start/end)
and only for the relevant steps (previous/next?).

Fixes #1055