chore: add ExceptionMapper to handle unhandeled...
... Excepions

This adds an ExceptionMapper to log the exception and return a JSON
response similar to the responses that the Syndesis REST provides.

User message is set to the message of the root exception with the
hypothesis that this exception's message is the the most appropriate.

There is an option to log the request details if the logging level for
`io.syndesis.verifier` is set to DEBUG.

Fixes #43