refactor: make dynamic action properties with e...
...num values render as HTML select

This detects that a metadata for a property contains `enum` and based on
that renders the property as a HTML select element. This would render
`displayValue` (or `label`) to the user instead of the `value` which
could/should hold the technical value, not meant to be shown to the
user.

The big issue is how the backend expects properties to be handed over at
a particular step in time. Now a property that is defined on the step
the user is in, and any subsequent step is not sent to the backend,
making it the same as if the user first arrived at the current step.

Fixes syndesisio/syndesis-verifier#45
Fixes syndesisio/syndesis-ui#1088
Fixes syndesisio/syndesis-ux#40
Fixes syndesisio/syndesis-rest#599