adding exception handling for OAuth issues (WIP)
Best effort error parsing

refactor: OAuth error handling

I wanted to share code paths for XML and JSON error responses to ease
adding other error message extraction paths. This changes to use Jackson
for both JSON and XML responses.
Could not find any usage of ErrorMap.request so I refactored the
ErrorMap to be a utility class that just extracts the error message from
the response.
It could happen that some of the responses start with whitespace
characters so I switched to using regexp to guess the format.
I thought that it would make sense if the error message is not parsed we
could benefit from having the response included verbatim, I'm thinking
situations like screenshots where even if the error is quite ugly to
present it would help us troubleshoot.