feat: Extended migrate-images.pl
and migrated the script itself to Perl (yes, P.E.R.L !).
For complex data mappings Perl is much better suited and
to adapt if for new versions should be easy. Just have a look
into the code and follow the comments.

Call it with `migrate-images.pl --source=brew` to pull from the brew registry,
`migrate-images.pl --source=pipeline` for the pipelines builds. "brew" is the default.