refactor: cleanup integration model
The only functional difference is addition of `@SkipNulls` annotation
that filters out SimpleStep.configuredProperties that contain `null`
values. Issue that manifests itself when trying to add a mapping step to
an integration and having no mapping defined at that point, in JSON it
looks like:

    {
        "configuredProperties": {
            "atlasmapping": ""
        }
    }

That is converted by `StringTrimmingJsonDeserializer` to
`(atlasmapping, null)`, and in trun results in a NullPointerException
when trying to deserialize the JSON.

Also adds a few `@Override` annotations, comments to empty blocks.