fix: ci locking
Uses `~` as delimiter in `project_lock_data()` to be certain that the
fields are parsed according to their positions, as using ` ` would lead
to fields being parsed out of order.

Sets the project/namespace annotation on the secret used for locking to
make sure that the `newallocator` check passes to confirm that the lock
has been taken.