feat: disable cache when debugging
Since there is little or no way to influence Infinispan cache from
outside of the running backend, when developing direct changes to
database will not be refleted until the `syndesis-rest` pod is
restarted.

With this if `JAVA_DEBUG` environment variable is set to `true` maximum
size of Infinispan will be set to `0` -- effectively disabling the
cache. Now changes direct to the database will be reflected in the
responses of the REST API.