fix(openshift): look at `availableReplicas` ins...
...tead of `readyReplicas`

Seems that we should be looking at `availableReplicas` instead of
`readyReplicas` to determine if the integration pod is running.

Looking at kubernetes reference[1] seems more appropriate for our use
case.

Tested on minishift with integration pod that was happy to run and a
integration that was not so happy to run, seems to reflect what we want
to achieve.

[1] https://v1-8.docs.kubernetes.io/docs/api-reference/v1.8/#deploymentstatus-v1beta2-apps