fix(ui): stop ng-chart from initiating change storm
I traced the ever-repainting kebab menu to chart component triggering
change event, seems that ng-chart is adding functions to array and that
in change detection triggers repaint of other directives.

I know I explain this wrong, I'm a Angular newb, though it seems to
work with this change. Perhaps there is a better way of doing this, not
sure.

Fixes #733