chore(rest): don't fail on `null` credentials
Currently oauth-proxy is not forwarding OpenShift OAuth token in the
`X-Forwarded-Access-Token` header if the API is invoked using Bearer
token. When `Authentication` with `null` credentials is passed on to
the `PreAuthenticatedAuthenticationProvider` it fails to authenticate
the request.

This adds a fixed credentials `"<missing>"` so that the authentication
provider can authenticate and pass the request through.

The drawback of this is that the `Authentication` stored in the
`SecurityContext` will not have access to OpenShift OAuth token (it will
be `"<missing>"`), and there is no way of fixing that unless we modify
oauth-proxy to include it for requests with Bearer token.