chore: configure oauth-proxy to allow bearer to...
...ken authentication

We need to allow non UI clients to access our backend APIs as for some
of them it is cumbersome to use OAuth flow and Cookies. This is
difficult for our IDE integration to do, for example.

oauth-proxy can be configured to also check if the `Authorization`
header of the incoming request contains an OAuth bearer token and check
if that token is valid to pass on the request to the API.

This configures oauth-proxy when running a _restricted_ variant of the
templates to that end.

Caveat is that in this configuration oauth-proxy is not pasing the
`X-Forwarded-Access-Token` header, and we will loose OpenShift OAuth
token that we require (for some endpoints).

[1] https://github.com/syndesisio/syndesis/issues/99#issuecomment-348017381