fix(rest): don't loose `type` property of JSON schema
When serializing `SyndesisMetadata<ObjectSchema>` `type` property of
the schema (i.e. `"type": "object") is lost.

This workarounds it by converting `SyndesisMetadata` to a
`Map<String, Object>` before serialization so that property is not lost.

Probably has something to do with generics and the way JAX-RS support
for Jackson insists on using the method return type as the type of
object to serialize.

Fixes #1295