fix(ui): data shape selection for filter step
When we need to decide what shape to use for the filter step the logic
picked the input data shape of the next step even if the mapping step
was not present in the integration.

This simplifies the selection of data shape for filter step by having
only one data shape to chose from and performing the selection in the
`IntegrationStepConfigureComponent`.

Hopefully the logic is a bit more straightforward now.