fix(rest): make sure that we don't overwrite `j...
...ava` data shape kinds

We don't want to overwrite `java` data shape kinds, the verifier can
return the shape of an object of a different kind, for instance it
introspects the service and returns a JSON schema (Salesforce does
this).
But we don't want to change shapes that the connector has defined as
kind `java`, the connector will in that case always set a Java object on
the body.