feat(server): hystrix and error handling for me...
...ta access

This adds the hystrix command to when performing the call to
syndesis-meta to fetch the dynamic metadata.

The command is configured in such a way that it provides the default or
no metadata instead of an error, if the syndesis-meta returns an error
or the request times out.

Any error or timeout will be reported using the new `_meta` feature.
This way the UI can work with the default, non dynamic, metadata, and
still present the error to the user.

Fixes #1659