fix(server): try to use correct input/output shapes
With this if the result of metadata invocation for data shapes that
should contain specification/type, i.e. all that don't have their kind
defined as `none` or `any, and do not in fact contain those, for
instance if the meta data invocation fails or returns an erroneous data
shape, the resulting data shape will be set to `kind=any`.

Fixes #1967